fix(cookie): remove domain hash prefix
In https://github.com/rneatherway/gh-slack/issues/71 it was discussed
and discovered that some versions of Slack now include a SHA256 of the
host name as a prefix of the cookie.

This makes the cookie value invalid as-is, and the cookie value needs to
be trimmed from the hash before being used.

Using `-tag network` I was able to run the test locally and confirmed
that this now works.

Fix https://github.com/rneatherway/gh-slack/issues/71
cc https://github.com/nobe4/gh-slack/pull/1
cc https://github.com/rneatherway/slack/pull/5